+++ R/liana_utils.R
@@ -1,3 +1,5 @@
+
+
 #' Function to translate Human genes to organism X orthologues
 #'
 #' @param op_resource resource in the format of OmniPath/LIANA